# openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "YOLO & SAM Inference Backend",
		"version": "0.1.0"
	},
	"paths": {
...
		"/yolo-models": {
			"get": {
				"tags": ["yolo"],
				"summary": "List Yolo Models",
				"operationId": "list_yolo_models_yolo_models_get",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/YoloModelInfo"
									},
									"type": "array",
									"title": "Response List Yolo Models Yolo Models Get"
								}
							}
						}
					}
				}
			},
			"post": {
				"tags": ["yolo"],
				"summary": "Upload Yolo Model",
				"operationId": "upload_yolo_model_yolo_models_post",
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"$ref": "#/components/schemas/Body_upload_yolo_model_yolo_models_post"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UploadModelResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/yolo-models/archive": {
			"post": {
				"tags": ["yolo"],
				"summary": "Upload Yolo Model Archive",
				"operationId": "upload_yolo_model_archive_yolo_models_archive_post",
				"requestBody": {
					"content": {
						"multipart/form-data": {
							"schema": {
								"$ref": "#/components/schemas/Body_upload_yolo_model_archive_yolo_models_archive_post"
							}
						}
					},
					"required": true
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UploadModelResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/yolo-models/{model_name}/weights": {
			"get": {
				"tags": ["yolo"],
				"summary": "Download Yolo Model Weights",
				"operationId": "download_yolo_model_weights_yolo_models__model_name__weights_get",
				"parameters": [
					{
						"name": "model_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Model Name"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/yolo-models/{model_name}/onnx": {
			"get": {
				"tags": ["yolo"],
				"summary": "Export Yolo Model Onnx",
				"operationId": "export_yolo_model_onnx_yolo_models__model_name__onnx_get",
				"parameters": [
					{
						"name": "model_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Model Name"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/yolo-models/{model_name}/classes": {
			"get": {
				"tags": ["yolo"],
				"summary": "Download Yolo Model Classes",
				"operationId": "download_yolo_model_classes_yolo_models__model_name__classes_get",
				"parameters": [
					{
						"name": "model_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Model Name"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/yolo-models/{model_name}": {
			"delete": {
				"tags": ["yolo"],
				"summary": "Delete Yolo Model",
				"operationId": "delete_yolo_model_yolo_models__model_name__delete",
				"parameters": [
					{
						"name": "model_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Model Name"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Successful Response"
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/yolo-models/{model_name}/rename": {
			"put": {
				"tags": ["yolo"],
				"summary": "Rename Yolo Model",
				"operationId": "rename_yolo_model_yolo_models__model_name__rename_put",
				"parameters": [
					{
						"name": "model_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Model Name"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RenameModelRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/YoloModelInfo"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/yolo-models/{model_name}/annotate": {
			"post": {
				"tags": ["yolo"],
				"summary": "Auto Annotate Images",
				"operationId": "auto_annotate_images_yolo_models__model_name__annotate_post",
				"parameters": [
					{
						"name": "model_name",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Model Name"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AutoAnnotateRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AutoAnnotateResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"AutoAnnotateRequest": {
				"properties": {
					"image_urls": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Image Urls"
					},
					"conf_threshold": {
						"anyOf": [
							{
								"type": "number"
							},
							{
								"type": "null"
							}
						],
						"title": "Conf Threshold",
						"default": 0.25
					},
					"imgsz": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Imgsz"
					},
					"class_map": {
						"anyOf": [
							{
								"additionalProperties": {
									"type": "string"
								},
								"type": "object"
							},
							{
								"type": "null"
							}
						],
						"title": "Class Map"
					}
				},
				"type": "object",
				"required": ["image_urls"],
				"title": "AutoAnnotateRequest"
			},
			"AutoAnnotateResponse": {
				"properties": {
					"annotations": {
						"items": {
							"items": {
								"additionalProperties": true,
								"type": "object"
							},
							"type": "array"
						},
						"type": "array",
						"title": "Annotations"
					}
				},
				"type": "object",
				"required": ["annotations"],
				"title": "AutoAnnotateResponse"
			},
...
			"Body_upload_yolo_model_archive_yolo_models_archive_post": {
				"properties": {
					"name": {
						"type": "string",
						"title": "Name"
					},
					"archive_file": {
						"type": "string",
						"format": "binary",
						"title": "Archive File"
					}
				},
				"type": "object",
				"required": ["name", "archive_file"],
				"title": "Body_upload_yolo_model_archive_yolo_models_archive_post"
			},
			"Body_upload_yolo_model_yolo_models_post": {
				"properties": {
					"name": {
						"type": "string",
						"title": "Name"
					},
					"weights_file": {
						"type": "string",
						"format": "binary",
						"title": "Weights File"
					},
					"classes_file": {
						"anyOf": [
							{
								"type": "string",
								"format": "binary"
							},
							{
								"type": "null"
							}
						],
						"title": "Classes File"
					}
				},
				"type": "object",
				"required": ["name", "weights_file"],
				"title": "Body_upload_yolo_model_yolo_models_post"
			},
...
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
			"RenameModelRequest": {
				"properties": {
					"new_name": {
						"type": "string",
						"title": "New Name"
					}
				},
				"type": "object",
				"required": ["new_name"],
				"title": "RenameModelRequest"
			},
...
			"UploadModelResponse": {
				"properties": {
					"name": {
						"type": "string",
						"title": "Name"
					},
					"classes": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Classes"
					},
					"message": {
						"type": "string",
						"title": "Message"
					}
				},
				"type": "object",
				"required": ["name", "classes", "message"],
				"title": "UploadModelResponse"
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					}
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError"
			},
...
			"YoloModelInfo": {
				"properties": {
					"name": {
						"type": "string",
						"title": "Name"
					},
					"classes": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"title": "Classes"
					},
					"date_add": {
						"type": "string",
						"title": "Date Add"
					},
					"size_bytes": {
						"type": "integer",
						"title": "Size Bytes"
					},
					"imgsz": {
						"anyOf": [
							{
								"type": "integer"
							},
							{
								"type": "null"
							}
						],
						"title": "Imgsz"
					},
					"trained": {
						"type": "boolean",
						"title": "Trained",
						"default": false
					}
				},
				"type": "object",
				"required": ["name", "classes", "date_add", "size_bytes"],
				"title": "YoloModelInfo"
			}
		}
	}
}